backend/requirements.txt
fastapi
uvicorn
python-multipart
scikit-learn
tensorflow
torch
transformers
spacy
huggingface_hub
openai
google-genai
joblib
numpy
pandas
python-dotenv
